+++ examples/snapshot/__tests__/clock.test.js
@@ -5,8 +5,7 @@
 import renderer from 'react-test-renderer';
 import Clock from '../Clock';
 
-jest.useFakeTimers();
-Date.now = jest.fn(() => 1482363367071);
+jest.useFakeTimers().setSystemTime(1482363367071);
 
 it('renders correctly', () => {
   const tree = renderer.create(<Clock />).toJSON();