+++ packages/jest-circus/src/utils.ts
@@ -179,8 +179,8 @@ const _makeTimeoutMessage = (
   `Exceeded timeout of ${formatTime(timeout)} for a ${
     isHook ? 'hook' : 'test'
   }${
-    takesDoneCallback && ' while waiting for `done()` to be called'
-  }.\nUse jest.setTimeout(newTimeout) to increase the timeout value, if this is a long-running test.`;
+    takesDoneCallback ? ' while waiting for `done()` to be called' : ''
+  }.\nAdd a timeout value to this test to increase the timeout, if this is a long-running test. See https://jestjs.io/docs/api#testname-fn-timeout.`;
 
 // Global values can be overwritten by mocks or tests. We'll capture
 // the original values in the variables before we require any files.