+++ packages/jest-worker/src/base/BaseWorkerPool.ts
@@ -7,11 +7,13 @@
 
 import mergeStream = require('merge-stream');
 import {
+  CHILD_MESSAGE_CALL_SETUP,
   CHILD_MESSAGE_END,
   PoolExitResult,
   WorkerInterface,
   WorkerOptions,
   WorkerPoolOptions,
+  WorkerStates,
 } from '../types';
 
 // How long to wait for the child process to terminate
@@ -27,9 +29,11 @@ export default class BaseWorkerPool {
   private readonly _stdout: NodeJS.ReadableStream;
   protected readonly _options: WorkerPoolOptions;
   private readonly _workers: Array<WorkerInterface>;
+  private readonly _workerPath: string;
 
   constructor(workerPath: string, options: WorkerPoolOptions) {
     this._options = options;
+    this._workerPath = workerPath;
     this._workers = new Array(options.numWorkers);
 
     const stdout = mergeStream();
@@ -83,8 +87,49 @@ export default class BaseWorkerPool {
     return this._workers[workerId];
   }
 
+  restartWorkerIfShutDown(workerId: number): void {
+    if (this._workers[workerId].state === WorkerStates.SHUT_DOWN) {
+      const {forkOptions, maxRetries, resourceLimits, setupArgs} =
+        this._options;
+      const workerOptions: WorkerOptions = {
+        forkOptions,
+        idleMemoryLimit: this._options.idleMemoryLimit,
+        maxRetries,
+        resourceLimits,
+        setupArgs,
+        workerId,
+        workerPath: this._workerPath,
+      };
+      const worker = this.createWorker(workerOptions);
+      this._workers[workerId] = worker;
+    }
+  }
+
   createWorker(_workerOptions: WorkerOptions): WorkerInterface {
     throw Error('Missing method createWorker in WorkerPool');
+  }
+
+  async start(): Promise<void> {
+    await Promise.all(
+      this._workers.map(async worker => {
+        await worker.waitForWorkerReady();
+
+        await new Promise<void>((resolve, reject) => {
+          worker.send(
+            [CHILD_MESSAGE_CALL_SETUP],
+            emptyMethod,
+            error => {
+              if (error) {
+                reject(error);
+              } else {
+                resolve();
+              }
+            },
+            emptyMethod,
+          );
+        });
+      }),
+    );
   }
 
   async end(): Promise<PoolExitResult> {